Create page.html
This is the HTML file that goes together with the GS script in the same Google Scripts App (see ReadMe for details on how to create files and put them together)

Currently, the code asks 4 questions of the user and then asks them to push the "speak" button. You can increase or reduce the number of questions and/or change what the questions say as you wish. You can also easily change the language of the speech that it recognizes. If you are a beginner user, please see the README file as it contains a YouTube link that explains how to do this easily and step-by-step. 

For more advanced users: 
1. Change the title on line 4
2. Change the label text for questions on lines 13, 16, 19, 22; you can also change the placeholders (sample responses) on lines 14, 17, 20, 23.
3. Change the language of speech recognition on line 70 (see README for a link to a list of language codes)
4. If you add or remove questions, add or remove from lines 45-54, especially making sure to change what is sent in line 50. Then, also amend the line in the GS script to match.